[feature] Added initial logic to like/dislike posts and add comments.

+++ src/main/java/org/logger421/poster/repositiories/PostRepository.java
@@ -4,5 +4,5 @@ import org.logger421.poster.models.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Integer> {
-    Post getPostById(int id);
+    Post getPostById(long id);
 }
